% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_lin_up_log_down.R
\name{auc_lin_up_log_down}
\alias{auc_lin_up_log_down}
\title{Linear Up-Log Down Trapezoidal Area Under the Concentration-Time Curve (AUC) Calculation Method}
\usage{
auc_lin_up_log_down(
  conc = NULL,
  time = NULL,
  exflag = NULL,
  interpolate = NULL,
  extrapolate = NULL,
  model = NULL,
  dosing_type = NULL,
  dosing_interval = NULL,
  told = NULL,
  prev_told = NULL,
  prev_tau = NULL,
  last_crit_factor = NULL,
  kel = NULL,
  orig_conc = NULL,
  orig_time = NULL,
  orgtime = NULL,
  includeNA = FALSE
)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{exflag}{The exclude flag data (given in a numeric vector)}

\item{interpolate}{The value to determine whether to interpolate data points (given in a logical form)}

\item{extrapolate}{The value to determine whether to extrapolate data points (given in a logical form)}

\item{model}{The model specification (either 'M1', 'M2', 'M3', or 'M4')}

\item{dosing_type}{The dosing type specification (either 'SD' or 'SS')}

\item{dosing_interval}{The current dosing interval (numeric value)}

\item{told}{The time of last dose (given in a numeric value)}

\item{prev_told}{The time of last dose from previous interval (given in a numeric value)}

\item{prev_tau}{The time duration of previous dosing interval (given in a numeric value)}

\item{last_crit_factor}{The criteria value for last time acceptance criteria (numeric value)}

\item{kel}{The KEL value (given as a numeric)}

\item{orig_conc}{The original (full) concentration data (given in a numeric vector)}

\item{orig_time}{The original (full) time data (given in a numeric vector)}

\item{orgtime}{The original time value from the map data ('nominal' or 'actual')}

\item{includeNA}{The value that determines if NA values should be removed from concentration and time data (given in logical value)}
}
\description{
Linear trapezoidal while concentrations are increasing, and log trapezoidal while
concentrations are decreasing; the assessment is made on a step by step basis for
each portion of the profile i.e. t1 to t2. If Ci or Ci+1 is 0 then the linear trapezoidal
rule is used.
}
