% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aumc_lin.R
\name{aumc_lin}
\alias{aumc_lin}
\title{Linear Trapezoidal Area Under the First Moment Curve (AUMC) Calculation Method}
\usage{
aumc_lin(conc = NULL, time = NULL, exflag = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{exflag}{The exclude flag data (given in a numeric vector)}
}
\description{
The AUMC linear method is used for the entire profile.
}
