% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aumc_log.R
\name{aumc_log}
\alias{aumc_log}
\title{Log Trapezoidal Area Under the First Moment Curve (AUMC) Calculation Method}
\usage{
aumc_log(conc = NULL, time = NULL, exflag = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{exflag}{The exclude flag data (given in a numeric vector)}
}
\description{
The logarithmic method is used for the entire profile. If Ci or Ci+1 is 0 then the
linear trapezoidal rule is used
}
