% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c0.R
\name{c0}
\alias{c0}
\title{Initial concentration resulting immediately after an IV bolus administration.}
\usage{
c0(conc = NULL, time = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}
}
\description{
Initial concentration resulting immediately after an IV bolus administration.
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item C0: initial concentration
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
