% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cendinf.R
\name{cendinf}
\alias{cendinf}
\title{Observed concentration at the end of infusion}
\usage{
cendinf(conc = NULL, time = NULL, dof = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{dof}{The duration of infusion data (given as a numeric value)}
}
\description{
This function gets the observed concentration at the end of infusion. In the case of multiple dosing, CENDINFi
is obtained by inspection of the data during each dosing interval i.
}
\details{
The function will ensure that the concentration data and the time data provided are numeric vectors and
are also of the same length. \cr
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item CENDINF: observed concentration at the end of infusion
}
}

\examples{
##########
## Data ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   30    ##    0   ##   2.89   ##
##   30    ##    1   ##   2.49   ##
##   30    ##    2   ##   2.47   ##
###################################

#Data mentioned will be used for the following example

cendinf()
#Error in cendinf: 'conc' and 'time' vectors are NULL

conc_vector <- c(2.89, 2.49, 2.47)
time_vector <- c(0, 1, 2)

cendinf(conc = conc_vector, time = time_vector, dof = 1)
#2.49

############
## Data 2 ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##    31   ##    0   ##     0    ##
##    31   ##    1   ##     0    ##
##    31   ##    2   ##     0    ##
###################################

#Data mentioned will be used for the following example

conc_vector <- c(0, 0, 0)
time_vector <- c(0, 1, 2)

cendinf(conc = conc_vector, time = time_vector, dof = NA)
#NA

############
## Data 3 ##
#################################
##  SID  ##  TIME  ##   CONC   ##
#################################
##   32  ##    0   ##   1.19   ##
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##   1.34   ##
##   32  ## "None" ##   1.32   ##
#################################

#Data mentioned will be used for the following example

conc_vector <- c(1.19, 1.23, 1.34, 1.32)
time_vector <- c(0, 1, 2, "None")

cendinf(conc = conc_vector, time = time_vector)
#Error in cendinf: 'time' is not a numeric vector

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
