% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clast_i.R
\name{clast_i}
\alias{clast_i}
\title{Last Measurable Plasma Concentration}
\usage{
clast_i(data = NULL, map = NULL, flag = NULL, sid = list(),
  interval = 0)
}
\arguments{
\item{data}{The model data that was generated by \code{\link{model}}}

\item{sid}{The subject ID (either can be the id number or 'all' to get clast for all the subjects)}
}
\description{
This function gets the last measurable plasma concentration that is obtained by the 
inspection of the data.
}
\details{
If all the concentrations are 0's then CLAST will return 'ND' for Not Determined. 
Also you must provide a valid sid with respect to the data.
}
\section{Returns}{

\strong{Dataset} \cr 
\itemize{
 \item SID: subject identification number 
 \item CLAST: last measurable plasma concentration
}
}

\examples{
########## 
## Data ##
#################################
##  SID  ##  TIME  ##  RESULT  ##
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
##   31  ##    0   ##      0   ##
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
##   32  ##    0   ##   1.19   ##
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##   1.34   ##
##   32  ##    4   ##    BLQ   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

clast()   
#No data found!

clast(data)  
#Object not of class NCA

mod <- model("~/data.csv")  
#Creates an NCA object with data represented in 'data' above
clast(mod)  
#Please specify for which subject you want to get the CLAST for!

clast(mod, sid = "all")  
# SID  CLAST
#  30   2.47
#  31     ND
#  32   1.34

clast(mod, sid = 32)  
# SID  CLAST
#  32   1.34

clast(mod, sid = 10)  
Invaild subject ID! Subject ID not found in the data provided!

##############
## nca_data ##
##############

mod2 <- model(nca_data, sid = "Subject", time = "Ntime", result = "Conc", dose = "Dose")

clast(mod2, "all")
#   SID CLAST
#1    1  3.28
#2    2  0.90
#3    3  1.05
#4    4  1.15
#5    5  1.57
#6    6  0.92
#7    7  1.15
#8    8  1.25
#9    9  1.12
#10  10  2.42
#11  11  0.00
#12  12  1.17

}
\author{
\itemize{
 \item Kevin McConnell
}
}
