% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmin_i.R
\name{cmin_i}
\alias{cmin_i}
\title{Minimum Observed Concentration}
\usage{
cmin_i(data = NULL, map = NULL, flag = NULL, sid = list(),
  interval = 0)
}
\arguments{
\item{data}{The model data that was generated by \code{\link{model}}}

\item{sid}{The subject ID (either can be the id number or 'all' to get cmin for all the subjects)}

\item{interval}{The endpoint of an dosing interval (if not provided then it will use the entire time interval)}
}
\description{
This function gets the minimum observed concentration that is obtained by the inspection of the data.
It is designed as the first occurance within a dosing interval. In the case of multiple dosing, CMINi 
is obtained by inspection of the data during each dosing interval i.
}
\details{
If all the concentrations are 0's then CMINi will return 'ND' for Not Determined. 
Also the interval must be in the range of the times listed in the data. 
And also you must provide a valid sid with respect to the data.
}
\section{Returns}{

\strong{Dataset} \cr 
\itemize{
 \item SID: subject identification number
 \item CMIN: minimum observed concentration
 \item INTERVAL: endpoint of the dosing interval
}
}

\examples{
##########
## Data ##
#################################
##  SID  ##  TIME  ##  RESULT  ##
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
##   31  ##    0   ##      0   ##
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
##   32  ##    0   ##   1.19   ##
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##   1.34   ##
##   32  ##    4   ##   1.32   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

cmin_i()   
#No data found!

cmin_i(data)  
#Object not of class NCA

mod <- model("~/data.csv")  
#Creates an NCA object with data represented in 'data' above
cmin_i(mod)  
#Please specify for which subject you want to get the CMINi for!

cmin_i(mod, sid = "all")  
#"2.47", "ND", "1.19"

cmin_i(mod, sid = 31)  
#"ND"

cmin_i(mod, sid = 10)  
#Invaild subject ID! Subject ID not found in the data provided!

cmin_i(mod, sid = 31, interval = -1)
#Interval provided is less than zero!

cmin_i(mod, sid = 31, interval = 4)
#Interval provided is greater than maximum time!

cmin_i(mod, sid = 30, interval = 1)
#"2.498"

cmin_i(mod, sid = "all", interval = 4)
#Interval provided is greater than maximum time!

##############
## nca_data ##
##############

mod2 <- model(nca_data, sid = "Subject", time = "Ntime", result = "Conc", dose = "Dose")

#   SID CMIN INTERVAL
#1    1 0.74        0
#2    2 0.00        0
#3    3 0.00        0
#4    4 0.00        0
#5    5 0.00        0
#6    6 0.00        0
#7    7 0.15        0
#8    8 0.00        0
#9    9 0.00        0
#10  10 0.00        0
#11  11 0.00        0
#12  12 0.00        0

}
\author{
\itemize{
 \item Kevin McConnell
}
}
