% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dosing_intervals.R
\name{create_dosing_intervals}
\alias{create_dosing_intervals}
\title{create_dosing_intervals updates concentration data with dosing intervals}
\usage{
create_dosing_intervals(data, map, flag, maxdosingintervals)
}
\arguments{
\item{data}{The dataframe that contains the raw data}

\item{map}{The dataframe that contains the map data}

\item{flag}{The dataframe that contains the flag data}

\item{maxdosingintervals}{number of maximum dosing intervals}
}
\value{
data
}
\description{
create_dosing_intervals updates concentration data with dosing intevals values to align with
selected TIME, DOSINGTYPE, TOLDi and TAUi values.
}
\details{
that map$TIME variable either points to a valid variable in data or is nominal/actual selection.
map$TIME is either a value of "Nominal" or "Actual" or it points directly to a column/data field in the data dataframe.
if map$TIME=="Nominal" then map$NOMTIME/map$NOMTIMEU will point to the column/data field in the data dataframe for TIME
if map$TIME=="Actual"  then map$ACTTIME/map$ACTTIMEU will point to the column/data field in the data dataframe for TIME
if map$TIME=="Nominal"  then map$NOMTIME will be used to create the dosing intervals with respect to TOLDi and TAUi
if map$TIME=="Actual"  then map$NOMTIME will be used to create the dosing intervals with respect to TOLDi and TAUi
}
\section{Returns}{

updated concentration data
}

\examples{
#No appropriate examples

}
