% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctroughend.R
\name{ctroughend}
\alias{ctroughend}
\title{Concentrations at the end of the ith dose interval}
\usage{
ctroughend(conc = NULL, time = NULL, tau = NULL, told = NULL, ctold = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{tau}{The tau data (given in a vector form)}

\item{told}{The time of last dose (given in a numeric value)}

\item{ctold}{The concentration at the time of last dose (given in a numeric value)}
}
\description{
This function gets the concentrations at the end of the ith dosing interval, defined as the observed concentration 
at the time of nominal time TAU during multiple dosing.
}
\details{
If the concentration at the end of the ith dose interval is not provided then it returns NA. \cr
The function will ensure that the concentration data and the time data provided are numeric vectors and
are also of the same length. \cr
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item CTROUGHEND: concentrations at the end of ith dose interval
}
}

\examples{
##########
## Data ##
############################################
##  SDEID  ##  TIME  ##   CONC   ##  TAU  ##
############################################
##   30    ##    0   ##   2.89   ##   2   ##
##   30    ##    1   ##   2.49   ##   2   ##
##   30    ##    2   ##   2.47   ##   2   ##
############################################

#Data mentioned will be used for the following example

#ctroughend()
#Error in ctrough: 'conc' and 'time' vectors are NULL

conc_vector <- c(2.89, 2.49, 2.47)
time_vector <- c(0, 1, 2)
tau_value <- 2

ctroughend(conc = conc_vector, time = time_vector, tau = tau_value)
#2.47

############
## Data 2 ##
############################################
##  SDEID  ##  TIME  ##   CONC   ##  TAU  ##
############################################
##   31    ##    1   ##   2.23   ##   4   ##
##   31    ##    2   ##   2.34   ##   4   ##
##   31    ##    3   ##   2.68   ##   4   ##
############################################

#Data mentioned will be used for the following example

conc_vector <- c(2.23, 2.34, 2.68)
time_vector <- c(1, 2, 3)
tau_value <- 4

ctroughend(conc = conc_vector, time = time_vector, tau = tau_value)
#NA

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
