% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependent_parameters.R
\name{dependent_parameters}
\alias{dependent_parameters}
\title{Dependent Parameters}
\usage{
dependent_parameters(parameter, dl = create_dependency_list(), depth = TRUE)
}
\arguments{
\item{parameter}{Parameter Name}

\item{dl}{Dependency List}

\item{depth}{Logical Value to enable sub-dependent parameters}
}
\description{
This function will generate a list of dependent parameters based on the specified parameter\cr
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item List of dependent parameters
}
}

\examples{
#No appropriate examples

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
