% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dof.R
\name{dof}
\alias{dof}
\title{Duration of infusion for constant infusion model M3}
\usage{
dof(data, map, dof_name = "DOF1")
}
\arguments{
\item{data}{input concentration dataset}

\item{map}{Model Configuration Template (MCT)}
}
\description{
This function returns the duration of infusion
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item DOF: duration of infusion
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
