% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosec.R
\name{dosec}
\alias{dosec}
\title{DOSEC - DOSEADJ - Dose Scaled to Concentration Units}
\usage{
dosec(data = NULL, map = NULL, idose = NULL)
}
\arguments{
\item{data}{The dataframe that contains the raw data}

\item{map}{The dataframe that contains the map data}
}
\description{
DOSEC - DOSEADJ - Dose Scaled to Concentration Units
}
\details{
This function scales the DOSE to DOSEC to the same amount unit for concentration. \cr
This scaling is required to place DOSE on the same unit basis for concentration for computation of
CLFO, CLO, CLFP, CLFTAU, CLP, CLTAU, CLTAUi, VZFO, VZFP, VZFTAU, VZTAUi, VZO, VZP. \cr
Assumptions are that input Model Configuration Template (MCT/map) has been updated by update_mct_data and
unit_conversion is used for DOSE unit scaling. \cr
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item DOSEC: DOSE adjusted for concentration units
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{Thomas.G.Tensfeldt@pfizer.com}
}
}
