% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_c0.R
\name{est_c0}
\alias{est_c0}
\title{Estimated C0 concentration estimated for IV bolus administration}
\usage{
est_c0(conc = NULL, time = NULL, npts = 2, verbose = FALSE, returnall = TRUE)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{npts}{The # of data points to utilize in back extrapolation (Default: 2)}

\item{verbose}{If TRUE print additional messages}

\item{returnall}{If TRUE return a list with the time, conc, and resulting est_c0, otherwise return est_c0}
}
\description{
Estimated C0 concentration obtained from back-extrapolation of the first of several
concentration datapoints obtained immediately after an IV bolus administration.
}
\details{
Note: consider option to scan time,conc data until a negative slope is obtained to avoid positive slopes
if no negative slope can be obtained with this option, return NA as est_c0. Not implemented as of 2019-09-24.
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item est_C0: estimated initial concentration
}
}

\examples{
time <- c(0.00,0.25,0.50,1.00,1.50,2.00,3.00,5.00,8.00,12.00,24.00,48.00,72.00,96.00,120.00,168.00,336.00,504.00)
conc <- c(0.0,125.0,122.0,126.0,127.0,130.0,136.0,118.0,99.0,85.4,63.2,36.2,23.5,17.4,11.5,0.0,0.0,0.0)
est_c0(conc=conc, time=time)
warning: <est_c0.R>: Following 0/zero/NA valued concentration data points were removed from regression:

  time conc
1     0    0
16  168    0
17  336    0
18  504    0
(Intercept) 
128.0738

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer Inc}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
