% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_concentration.R
\name{estimate_concentration}
\alias{estimate_concentration}
\title{Estimate concentrations from results of linear regression of the time-conc data}
\usage{
estimate_concentration(time, conc, slope)
}
\arguments{
\item{conc}{The vector of concentration data}

\item{slope}{The slope of the log-lin regression of the time-conc data generated following selection of the data points}

\item{tine}{The vector of time point data}
}
\description{
Estimate concentrations from results of linear regression of the time-conc data
}
\author{
\itemize{
 \item \strong{Tom Tensfeldt, Pfizer}
 \item website: \url{}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
