% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot.R
\name{generate_plot}
\alias{generate_plot}
\title{Generate Plot}
\usage{
generate_plot(data = NULL, x = NULL, y = NULL, groupBy = NULL,
  groupMethod = NULL, groupType = list(), arrangeBy = NULL,
  pageBy = NULL, colorBy = NULL, plotType = list(), yOrient = FALSE,
  title = NULL, xLab = NULL, yLab = NULL, footnote = NULL,
  legendTitle = NULL, outputFileName = NULL, oriented = FALSE)
}
\arguments{
\item{data}{The data frame object with the required data}

\item{x}{The column that will be used for the x-coordinate}

\item{y}{The column that will be used for the y-coordinate}

\item{groupBy}{A list of columns that will be used to group the dataset by}

\item{groupMethod}{A method by which the group by calculation should occur ('mean' or 'median')}

\item{groupType}{A list of the type of group by to perform ('line' and/or 'point')}

\item{arrangeBy}{The column by which the dataset will be arranged or combined by}

\item{pageBy}{A list of columns by which the dataset will be paged by}

\item{colorBy}{A column by which the dataset will be colored by}

\item{plotType}{A list of the type of plot to display ('line' and/or 'point')}

\item{yOrient}{A boolean value that is used to change the orientation of the text of y-axis}

\item{title}{The title of the plot}

\item{xLab}{The label for the x-axis}

\item{yLab}{The label for the y-axis}

\item{footnote}{A string note that will be appended at the bottom of the plot in the PDF document}

\item{outputFileName}{The file name of the output file (include '.pdf')}

\item{oriented}{A boolean value that is used to determine whether PDF will output in portrait or landscape}
}
\description{
This function reads in a dataframe and creates a PDF document generating a plot.
}
\details{
\itemize{
 \item 'data', 'x', 'y' and 'arrangeBy' by defauly are 'NULL'
 \item 'groupBy', 'pageBy' by default are an empty list ('list()')
 \item 'groupMethod' by default is 'NULL', but can be set to either 'mean' or 'median'
 \item 'groupType' and 'plotType by default are 'NULL', but can be either 'line' or 'point'
 \item 'colorBy', 'title', 'xLab', 'yLab', 'footnote' by default are 'NULL'
 \item 'yOrient' by default is 'FALSE' and will result in normal tickmark text
 \item 'outputFileName' by defualt is 'NULL' and will take the name of 'generate-plot.pdf'
 \item 'oriented' by default is 'FALSE' and will result in portrait
 \item Either 'groupType' or 'dataType' needs to be provided to generate the plot.
 \item If 'groupBy' is specified then both 'groupMethod' and 'groupType' also needs to be specified.
}
}
\note{
Erros will be throw if invalid input for 'data', 'x', 'y', 'plotType', 'oriented' is passed.
}
\section{Returns}{

\strong{PDF} \cr 
\itemize{
 \item Plot generated with specified options 
}
}

\examples{

#Arguments
Data = plot_dataframe
X = "PKPTMS"
Y = "PKCNC"
GroupBy = "TREATXT"
GroupMethod = "mean"
GroupType = list("line", "point")
PlotType = list("line")
XLab = "Time"
YLab = "Concentration (mg/ml)"
Title = "Conc. vs Time"
Footnote = "This is a test footnote"
YOrient = TRUE
PageBy = "SUBJID"
LegendTitle = "Treatment"
Oriented = TRUE

#TEST CASE ONE

#Function call
generate_plot(data = Data,
  x = X,
  y = Y, 
  groupBy = GroupBy, 
  groupMethod = GroupMethod, 
  groupType = GroupType, 
  plotType = PlotType, 
  xLab = XLab, 
  yLab = YLab, 
  title = Title, 
  footnote = Footnote, 
  yOrient = YOrient, 
  pageBy = PageBy, 
  legendTitle = LegendTitle, 
  oriented = Oriented
)

Will add more test cases with outputs soon

}
\author{
\itemize{
 \item Kevin McConnell \email{kmcconnell@rudrya.com}
}
}
