% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_table.R
\name{generate_table}
\alias{generate_table}
\title{Generate Table}
\usage{
generate_table(path = NULL, title = NULL, pageNumber = TRUE,
  dateGeneration = TRUE, numOfRows = 40, oriented = FALSE,
  groupBy = list(), pageBy = list(), orderBy = list(),
  columnType = list(), columnToDisplay = list(),
  columnNamesToDisplay = list(), footnote = list(), tableScale = 1,
  tableResize = 1, leftMargin = 0)
}
\arguments{
\item{path}{The path of the XML data file}

\item{title}{The title of the table that will appear in the PDF document}

\item{pageNumber}{A boolean value used to display page numbers}

\item{dateGeneration}{A boolean value used to generate date}

\item{numOfRows}{A numeric value that is used to determine the number of rows in each page}

\item{oriented}{A boolean value that is used to determine whether PDF will output in portrait or landscape}

\item{groupBy}{A list of columns by which the dataset will be grouped by}

\item{pageBy}{A list of columns by which the dataset will be paged by}

\item{orderBy}{A list of columns that will be used to order the dataset}

\item{columnType}{A list of columns names mappings which will be used to select either character or numeric for certain columns, by default this behavior will be auto-generated}

\item{columnToDisplay}{A list of columns to display, by default all are displayed (specify the order as well)}

\item{columnNamesToDisplay}{A list of columns names mappings which will be used instead of the column names of the dataset}

\item{footnote}{A list of notes that will be appended at the bottom of the table in the PDF document}

\item{tableScale}{A numeric value that will be used to scale the table outputted in the PDF document}

\item{tableResize}{A numeric value that will be used to resize the table outputted in the PDF document}

\item{leftMargin}{A numeric value that will be used to scale the page's left margin in the PDF document (in inches)}

\item{outputFileName}{The file name of the output file (include '.pdf')}
}
\description{
This function reads in a dataframe and creates a PDF document generating
a latex table.
}
\details{
\itemize{
 \item 'path' and 'title' by defauly are 'NULL'
 \item 'groupBy', 'pageBy', 'orderBy', 'columnToDisplay', 'columnNamesToDisplay', 'columnType' and 'footnote' by default are an empty list ('list()')
 \item 'pageNumber' and 'dateGeneration' by defualt are 'TRUE'
 \item 'outputFileName' by defualt is 'NULL' and will take the name of 'huxtable-output.pdf'
 \item 'oriented' by default is 'FALSE' and will result in portrait
 \item 'numOfRows' by default is set to '40'
 \item 'tableScale' by default is set to '1'
 \item 'tableResize' by default is set to '1'
 \item 'leftMargin' by default is set to '0'
 \item Both 'tableScale' and 'tableResize' tend to perform similar task but they work better together
 \item Both 'tableScale' and 'tableResize' values are multiplied to a 'textwidth' value of 390pt (1pt = 0.3515mm) to determine the final value
}
}
\note{
Erros will be throw if invalid input for 'path', 'numOfRows', 'tableScale', 'tableResize', 'leftMargin' is passed.
}
\section{Returns}{

\strong{PDF} \cr 
\itemize{
 \item Table generated with specified options
}
}

\examples{

#TEST CASE ONE

#Arguments
Path = "/eNCA/tc00/M1SD/B1261002.in"
GroupBy = list('SUBJID', 'TREATXT')                 
PageBy = list()                  
OrderBy = list('SUBJID')              
ColumnType = list(SUBJID = "character", PKDATAROWID = "character", PERIOD = "numeric")            
ColumnToDisplay = list("SUBJID", "TREATXT", "PERIOD", "DOSE", "PKDATAROWID", "PKPTMR", "PKPTM", "PKCNC")        
ColumnNamesToDisplay = list(SUBJID = "Subject ID", TREATXT = "Treatment Group", PKCNC = "Concentration (NG/ML)")  
Footnote = list("this is a test", "this is a second test")              
Title = "PKL001"
PageNumber = TRUE
OutputFileName = NULL
DateGeneration = TRUE
NumOfRows = 40

#Function call
generate_table(
  path = Path, 
  title = Title, 
  pageNumber = PageNumber, 
  dateGeneration = DateGeneration, 
  numOfRows = NumOfRows, 
  groupBy = GroupBy, 
  pageBy = PageBy, 
  orderBy = OrderBy, 
  columnType = ColumnType, 
  columnToDisplay = ColumnToDisplay, 
  columnNamesToDisplay = ColumnNamesToDisplay, 
  footnote = Footnote
) 

Will add more test cases with outputs soon

}
\author{
\itemize{
 \item Kevin McConnell \email{kmcconnell@rudrya.com}
}
}
