% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasttime.R
\name{lasttime}
\alias{lasttime}
\title{Time at Last Measurable Concentration}
\usage{
lasttime(conc = NULL, time = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}
}
\description{
This is the time of the last measured concentration point of the profile INCLUDING any points at 
or below the Lower limit of Quantitation, BLQ’s . This observation can be later than TLAST.. \cr
}
\details{
LASTTIME serves, in one use case as a criteria for acceptability or quality flags to ensure 
there’s sufficient data to report certain parameters, for example, AUCTAU. If AUCTAU not of 
sufficient quality, then analyst would also not likely report CL. \cr
}
\section{Returns}{

\strong{Value} \cr 
\itemize{
 \item LASTTIME: time at last measureable concentration
}
}

\examples{
##########
## Data ##
#################################
##  SID  ##  TIME  ##   CONC   ## 
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

tlast()   
#Error in tlast: 'conc' and 'time' vectors are NULL

conc_vector <- data$CONC
time_vector <- data$TIME

tlast(conc = conc_vector, time = time_vector)
#2
 
############
## Data 2 ##
#################################
##  SID  ##  TIME  ##   CONC   ## 
#################################
##   31  ##    0   ##      0   ## 
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
#################################

data2 <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

conc_vector <- data2$CONC
time_vector <- data2$TIME

tlast(conc = conc_vector, time = time_vector)
#NA

############
## Data 3 ##
#################################
##  SID  ##  TIME  ##   CONC   ## 
#################################
##   32  ##    0   ##   1.19   ## 
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##   1.34   ##
##   32  ##    3   ##   BLQ    ##
#################################

data3 <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

conc_vector <- data3$CONC
time_vector <- data3$TIME

tlast(conc = conc_vector, time = time_vector)
#2

}
\author{
\itemize{
 \item Kevin McConnell
}
}
