% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenumeric.R
\name{makenumeric}
\alias{makenumeric}
\title{makenumeric converts/forces individual vectors to numeric values}
\usage{
makenumeric(x)
}
\arguments{
\item{x}{vector to force to numeric}
}
\value{
numeric vector

#' @author
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
\description{
makenumeric attempts to force a vector's values to all numeric values.
}
