% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxrate.R
\name{maxrate}
\alias{maxrate}
\title{Maximum observed excretion rate}
\usage{
maxrate(rate = NULL)
}
\arguments{
\item{rate}{The rate data (given in a vector form)}
}
\description{
Maximum observed excretion rate
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item MAXRATE: maximum rate
}
}

\examples{
#No appropriate examples

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
