% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midptlast.R
\name{midptlast}
\alias{midptlast}
\title{Midpoint of collection interval associated with last measurable (nonzero) rate}
\usage{
midptlast(midpt = NULL, rate = NULL)
}
\arguments{
\item{midpt}{The midpoint data (given in a vector form)}

\item{rate}{The rate data (given in a vector form)}
}
\description{
Midpoint of collection interval associated with last measurable (nonzero) rate
}
\section{Returns}{

\strong{Value} \cr 
\itemize{
 \item MIDPTLAST: last measurable rate
}
}

\examples{
#No appropriate examples

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
