% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{model_parameters}
\alias{model_parameters}
\title{model_parameters}
\usage{
model_parameters(model, dl = create_dependency_list())
}
\arguments{
\item{model}{Model and Dosing Type}

\item{dl}{Dependency List (Please refer to \code{\link{create_dependency_list}})}
}
\description{
This function will generate a list of parameters based on the specified model and dosing type\cr
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item List of parameters based on model and dosing type
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
