% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_spec.R
\name{model_spec}
\alias{model_spec}
\title{model_spec generates a brief description of the Model Configuration Template provided}
\usage{
model_spec(map)
}
\arguments{
\item{map}{Model Configuration Template (MCT)}
}
\description{
Generates a description string containing the MODEL and DOSINGTYPE from the
Model Configuration Template (MCT)
}
\section{Returns}{

\strong{TRUE} \cr
OR \cr
\strong{FALSE} \cr
}

\examples{
No appropriate examples

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
