% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opennca_version.R
\name{opennca_version}
\alias{opennca_version}
\title{Version of openNCA}
\usage{
opennca_version(print_version = TRUE)
}
\arguments{
\item{print_version}{Logical value that determines if the version numbers prints to the console}
}
\description{
This function prints and returns the version of the openNCA R Package
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item openNCA Version: X.X.X
}
}

\examples{
opennca_version()
#sopenNCA computation engine version <X.X.X>

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
