% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_indices.R
\name{parameter_indices}
\alias{parameter_indices}
\title{parameter_indices}
\usage{
parameter_indices(parameter_names, parameter_list, simplify = FALSE)
}
\arguments{
\item{parameter_names}{The character value, optionally vector of character values, of the regular expression(s) to search in the parameter_list}

\item{parameter_list}{The character vector of parameter names from the PARAMETERLIST of the Model Configuration Template}

\item{simplify}{If simplify is TRUE, default = TRUE, then return a single vector of indices. If FALSE, return a list of indices, named by entries in parameter_name}
}
\description{
This function returns the locations of parameter name values in the provided parameter_list character vector
}
\section{Returns}{

\strong{TRUE} \cr
OR \cr
\strong{FALSE} \cr
}

\examples{
plist <- c("AUC", "RATEN", "RATE1", "abc", "TMAX", "RATE3", "RATEa", "defg")

parameter_indices("^RATEN$", plist, simplify=FALSE)
#$RATEN
#[1] 2
#
parameter_indices("^RATEN$", plist, simplify=TRUE)
#RATEN 
#2

parameter_indices("^(RATE)([0-9]*?|A|N)$", plist, simplify=FALSE)
#$RATEN
#[1] 2
#
#$RATE1
#[1] 3
#
#$RATE3
#[1] 6
#
#$RATEa
#[1] 7

parameter_indices("^(RATE)([0-9]*?|A|N)$", plist, simplify=TRUE)
#RATEN RATE1 RATE3 RATEa 
#    2     3     6     7 

parameter_indices(c("^(RATE)([0-9]*?|A|N)$", "TMAX", "AUC"), plist, simplify=FALSE)
#$RATEN
#[1] 2
#
#$RATE1
#[1] 3
#
#$RATE3
#[1] 6
#
#$RATEa
#[1] 7
#
#$TMAX
#[1] 5
#
#$AUC
#[1] 1
#
#parameter_indices(c("^(RATE)([0-9]*?|A|N)$", "TMAX", "AUC"), plist, simplify=TRUE)
#RATEN RATE1 RATE3 RATEa  TMAX   AUC 
#    2     3     6     7     5     1 

parameter_indices(c("qrx"), plist, simplify=FALSE)
#named list()

parameter_indices("qrx", plist, simplify=FALSE)
#named list()

parameter_indices("qrx", plist, simplify=TRUE)
#[1] FALSE

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
