% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_regex.R
\name{parameter_regex}
\alias{parameter_regex}
\title{parameter_regex}
\usage{
parameter_regex(parameter, dl = create_dependency_list())
}
\arguments{
\item{parameter}{Parameter}

\item{dl}{Dependency List (Please refer to \code{\link{create_dependency_list}})}
}
\description{
This function returns the regular expression associated to the specified parameter
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item Regular Expresssion
}
}

\examples{


}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
