% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_by_class.R
\name{parameters_by_class}
\alias{parameters_by_class}
\title{parameter_by_class}
\usage{
parameters_by_class(unitclass, checklist, dl = create_dependency_list())
}
\arguments{
\item{unitclass}{Unit Class}

\item{checklist}{The character vector of parameter names from the PARAMETERLIST of the Model Configuration Template}

\item{dl}{Dependency List (Please refer to \code{\link{create_dependency_list}})}
}
\description{
This function returns the list of parameters based on the each Unit Class
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item List of Parameters based on each Unit Class
}
}

\examples{


}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
