% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.reg.R
\name{parse.reg}
\alias{parse.reg}
\title{parse.reg}
\usage{
parse.reg(res, result)
}
\arguments{
\item{res}{character string vector}

\item{result}{value from regexpr parse eval}
}
\value{
dataframe with results of the captures, with named fields if possible
}
\description{
for a single un-named capture field from a regular expression
 return the character string representing the captured value.
parse regular expression result
}
\examples{
parameters <- c("AUC", "THALF", "CMAX", "CMAXi", "TAU1", "TOLD1", "AUCINFP", "TAU2", "TAU", "TOLD2")
k <- regexpr("(?<basename>.+?)(?<index>i|[0-9]+)?$", parameters, ignore.case=FALSE, perl=TRUE)
terms <- parse.reg(parameters, k)
#> print(terms)
#   basename index
#1       AUC      
#2     THALF      
#3      CMAX      
#4      CMAX     i
#5       TAU     1
#6      TOLD     1
#7   AUCINFP      
#8       TAU     2
#9       TAU      
#10     TOLD     2

}
