% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{rate}
\alias{rate}
\title{Excretion Rate for each collection interval}
\usage{
rate(
  start_time = NULL,
  end_time = NULL,
  conc = NULL,
  vol = NULL,
  volu = NULL,
  type = "INTERVAL",
  map = NULL
)
}
\arguments{
\item{start_time}{The start time data (given in a vector form)}

\item{end_time}{The end time data (given in a vector form)}

\item{conc}{The concentration data (given in a vector form)}

\item{vol}{The volume data (given in a vector form)}

\item{volu}{The volume units data (given in a vector form)}

\item{type}{The type associated to rate (usually 'INTERVAL')}

\item{map}{The dataframe that contains the map data}
}
\description{
Excretion Rate for each collection interval
}
