% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratelast.R
\name{ratelast}
\alias{ratelast}
\title{Last measurable (nonzero) rate}
\usage{
ratelast(rate = NULL)
}
\arguments{
\item{rate}{The rate data (given in a vector form)}
}
\description{
Last measurable (nonzero) rate
}
\details{
If all the RATE's are 0's then RATELAST will return 'NA'.
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item RATELAST: last measurable rate
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
