% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmempty.R
\name{rmempty}
\alias{rmempty}
\title{rmempty will remove all empty columns from a dataframe}
\usage{
rmempty(d, pattern, quiet = FALSE)
}
\arguments{
\item{d}{dataframe to remove empty columns}

\item{pattern}{only remove columns that match pattern}

\item{quiet}{FALSE=display messages, TRUE=quiet operation}
}
\value{
dataframe reduced with empty columns removed

#' @author
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
\description{
Remove columns that are completely empty, i.e. either all values are NA or "",
from a dataframe
}
