% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_M1_SD_computation.R
\name{run_M1_SD_computation}
\alias{run_M1_SD_computation}
\title{Run M1 SD Computation}
\usage{
run_M1_SD_computation(data = NULL, map = NULL, method = 1, model = "M1",
  parameter = "SD", return = list())
}
\arguments{
\item{data}{The dataframe that contians the raw data}

\item{map}{The dataframe that contians the map data}

\item{method}{The AUC method to use}

\item{model}{This is the model type}

\item{parameter}{This is either single dose (SD) or steady state (SS)}

\item{return}{The list of parameters to return (by defualt it is empty, which means it will retunr all parameters)}
}
\description{
This function will compute all the relevant parameters for a M1 model Single Dose (SD).\cr
}
\details{
\strong{Methods:} You can use the following methods to calculate AUC: \cr
\enumerate{
 \item \strong{1: Linear-Log Trapazoidal Rule}(default method): The linear method is used up to Tmax (the 
 first occurance of Cmax) and the log trapezoidal method is used for the remainder of the profile. If
 Ci or Ci+1 is 0 then the linear trapezoidal rule is used.
 \item \strong{2: Linear Trapazoidal Rule}: The linear method is used for the entire profile.
 \item \strong{3: Log Trapazoidal Rule}: The log trapezoidal method is used for the entire profile. If
 Ci or Ci+1 is 0 then the linear trapezoidal rule is used.
 \item \strong{4: Linear Up - Log Down Trapazoidal Rule}: Linear trapezoidal while the concentrations
 are increasing and log trapezoidal while the concentration are decreasing, the assessment is made on
 a step basis for each portion of the profile i.e. t1 to t2. If Ci or Ci+1 is 0 then the linear 
 trapezoidal rule is used.
}
You can specify the options to subset the list of parameters that are returned: \cr
\strong{Return List options} \cr  
\enumerate{
 \item \strong{cmax}: Refer to \code{\link{cmax}} for more details
 \item \strong{cmin}: Refer to \code{\link{cmin}} for more details
 \item \strong{clast}: Refer to \code{\link{clast}} for more details
 \item \strong{cmax_c}: Refer to \code{\link{cmaxc}} for more details
 \item \strong{cmax_dn}: Refer to \code{\link{cmax_dn}} for more details
 \item \strong{tmax}: Refer to \code{\link{tmax}} for more details
 \item \strong{tmin}: Refer to \code{\link{tmin}} for more details
 \item \strong{tlast}: Refer to \code{\link{tlast}} for more details
 \item \strong{tlag}: Refer to \code{\link{tlag}} for more details
 \item \strong{kel}: Refer to \code{\link{kel}} for more details
 \item \strong{kelr}: Refer to \code{\link{kel_r}} for more details
 \item \strong{lasttime}: Refer to \code{\link{lasttime}} for more details
 \item \strong{auc_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{auc_all}: Refer to \code{\link{auc_all}} for more details
 \item \strong{auc_last}: Refer to \code{\link{auc_last}} for more details
 \item \strong{auc_last_c}: Refer to \code{\link{auc_lastc}} for more details
 \item \strong{auc_last_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{aumc_last}: Refer to \code{\link{aumc_last}} for more details
 \item \strong{auc_t1_t2}: Refer to \code{\link{auc_t1_t2}} for more details
 \item \strong{auc_inf_o}: Refer to \code{\link{auc_inf_o}} for more details
 \item \strong{auc_inf_o_c}: Refer to \code{\link{auc_inf_oc}} for more details
 \item \strong{auc_inf_p}: Refer to \code{\link{auc_inf_p}} for more details
 \item \strong{auc_inf_p_c}: Refer to \code{\link{auc_inf_pc}} for more details
 \item \strong{auc_inf_o_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{auc_inf_p_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{aumc_inf_o}: Refer to \code{\link{aumc_inf_o}} for more details
 \item \strong{aumc_inf_p}: Refer to \code{\link{aumc_inf_p}} for more details
 \item \strong{mrt_last}: Refer to \code{\link{mrt_last}} for more details
 \item \strong{mrto}: Refer to \code{\link{mrt_evif_o}} for more details
 \item \strong{mrtp}: Refer to \code{\link{mrt_evif_p}} for more details
 \item \strong{auc_xpct_o}: Refer to \code{\link{auc_XpctO}} for more details
 \item \strong{auc_xpct_p}: Refer to \code{\link{auc_XpctP}} for more details
 \item \strong{aumc_xpct_o}: Refer to \code{\link{aumc_XpctO}} for more details
 \item \strong{aumc_xpct_p}: Refer to \code{\link{aumc_XpctP}} for more details
 \item \strong{clo}: Refer to \code{\link{clo}} for more details
 \item \strong{clfo}: Refer to \code{\link{clfo}} for more details
 \item \strong{clfow}: Refer to \code{\link{clfow}} for more details
 \item \strong{clfp}: Refer to \code{\link{clfp}} for more details
 \item \strong{clfpw}: Refer to \code{\link{clfpw}} for more details
 \item \strong{vzfo}: Refer to \code{\link{vzfo}} for more details
 \item \strong{vzfow}: Refer to \code{\link{vzfow}} for more details
 \item \strong{vzfp}: Refer to \code{\link{vzfp}} for more details
 \item \strong{vzfpw}: Refer to \code{\link{vzfpw}} for more details
}
}
\section{Note}{

By default all the return list options are selected and calculated
}

\section{Returns}{

\strong{Dataset} \cr
}

\examples{
##########
## Data ##
########################################
##  SID  ##  TIME  ##  RESULT  ## ... ##
########################################
##   30  ##    0   ##   2.89   ##     ##
##   30  ##    1   ##   2.49   ##     ##
##   30  ##    2   ##   2.47   ##     ##
##   31  ##    0   ##      0   ##     ##
##   31  ##    1   ##   1.00   ##     ##
##   31  ##    2   ##      0   ##     ##
##   32  ##    0   ##   1.19   ##     ##
##   32  ##    1   ##   1.23   ##     ##
##   32  ##    2   ##   1.34   ##     ##
##   32  ##    4   ##   1.32   ##     ##
########################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

#########
## Map ## 
##################################################
##  PARAMETER  ##  MODEL  ##  DOSINGTYPE ## ... ##
##################################################
##    VALUE    ##    M1   ##     SD      ##     ##
##################################################

map <- data.frame(...)
#Same map as above, just represented as a dataframe

run_computation(data = data, map = map)
#Generates the M1 SD computation results

run_computation(data = data, map = map, flag = flag)
#Generates the M1 SD computation results

}
\author{
\itemize{
 \item Kevin McConnell
}
}
