% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_M2_SS_computation.R
\name{run_M2_SS_computation}
\alias{run_M2_SS_computation}
\title{Run M2 SS Computation}
\usage{
run_M2_SS_computation(data = NULL, map = NULL, method = 1,
  model_regex = "^M2(SS)*?$", parameter_list = list(),
  return_list = list())
}
\arguments{
\item{data}{The dataframe that contains the raw data}

\item{map}{The dataframe that contains the map data}

\item{method}{The methos of calculation for Area Under the Curve (see Details below)}

\item{parameter_list}{The list of parameters to calculate (empty by default)}

\item{return_list}{The list of parameters to return from the original data (empty by default)}

\item{display_list}{The list of parameters to return (empty by default)}
}
\description{
This function will compute all the relevant parameters for a M2 model Stedy State (SS).\cr
}
\details{
\strong{Methods:} You can use the following methods to calculate AUC: \cr
\enumerate{
 \item \strong{1: Linear-Log Trapazoidal Rule}(default method): The linear method is used up to Tmax (the
 first occurance of Cmax) and the log trapezoidal method is used for the remainder of the profile. If
 Ci or Ci+1 is 0 then the linear trapezoidal rule is used.
 \item \strong{2: Linear Trapazoidal Rule}: The linear method is used for the entire profile.
 \item \strong{3: Log Trapazoidal Rule}: The log trapezoidal method is used for the entire profile. If
 Ci or Ci+1 is 0 then the linear trapezoidal rule is used.
 \item \strong{4: Linear Up - Log Down Trapazoidal Rule}: Linear trapezoidal while the concentrations
 are increasing and log trapezoidal while the concentration are decreasing, the assessment is made on
 a step basis for each portion of the profile i.e. t1 to t2. If Ci or Ci+1 is 0 then the linear
 trapezoidal rule is used.
}
The following are the functions that this function uses to generate parameters: \cr
\strong{Additional Functions Used} \cr
\enumerate{
 \item \strong{c0}: Refer to \code{\link{c0}} for more details
 \item \strong{v0}: Refer to \code{\link{v0}} for more details
 \item \strong{cmax}: Refer to \code{\link{cmax}} for more details
 \item \strong{cmax_c}: Refer to \code{\link{cmaxc}} for more details
 \item \strong{cmax_dn}: Refer to \code{\link{cmax_dn}} for more details
 \item \strong{cmin}: Refer to \code{\link{cmin}} for more details
 \item \strong{clast}: Refer to \code{\link{clast}} for more details
 \item \strong{tmax}: Refer to \code{\link{tmax}} for more details
 \item \strong{tmin}: Refer to \code{\link{tmin}} for more details
 \item \strong{tlast}: Refer to \code{\link{tlast}} for more details
 \item \strong{kel}: Refer to \code{\link{kel}} for more details
 \item \strong{kelr}: Refer to \code{\link{kel_r}} for more details
 \item \strong{lasttime}: Refer to \code{\link{lasttime}} for more details
 \item \strong{auc_all}: Refer to \code{\link{auc_all}} for more details
 \item \strong{auc_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{auc_last}: Refer to \code{\link{auc_last}} for more details
 \item \strong{auc_last_c}: Refer to \code{\link{auc_lastc}} for more details
 \item \strong{auc_last_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{aumc_last}: Refer to \code{\link{aumc_last}} for more details
 \item \strong{auc_t1_t2}: Refer to \code{\link{auc_t1_t2}} for more details
 \item \strong{auc_inf_o}: Refer to \code{\link{auc_inf_o}} for more details
 \item \strong{auc_inf_o_c}: Refer to \code{\link{auc_inf_oc}} for more details
 \item \strong{auc_inf_o_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{cest}: Refer to \code{\link{cest}} for more details
 \item \strong{auc_inf_p}: Refer to \code{\link{auc_inf_p}} for more details
 \item \strong{auc_inf_p_c}: Refer to \code{\link{auc_inf_pc}} for more detail
 \item \strong{auc_inf_p_dn}: Refer to \code{\link{auc_dn}} for more details
 \item \strong{aumc_inf_o}: Refer to \code{\link{aumc_inf_o}} for more details
 \item \strong{aumc_inf_p}: Refer to \code{\link{aumc_inf_p}} for more details
 \item \strong{auc_tau}: Refer to \code{\link{auc_tau}} for more details
 \item \strong{aumc_tau}: Refer to \code{\link{aumc_tau}} for more details
 \item \strong{mrt_last}: Refer to \code{\link{mrt_last}} for more details
 \item \strong{mrt_ivif_o}: Refer to \code{\link{mrt_ivif_o}} for more details
 \item \strong{mrt_ivif_p}: Refer to \code{\link{mrt_ivif_p}} for more detail
 \item \strong{auc_XpctO}: Refer to \code{\link{auc_XpctO}} for more details
 \item \strong{auc_XpctP}: Refer to \code{\link{auc_XpctP}} for more details
 \item \strong{aumc_XpctO}: Refer to \code{\link{aumc_XpctO}} for more details
 \item \strong{aumc_XpctP}: Refer to \code{\link{aumc_XpctP}} for more details
 \item \strong{cav}: Refer to \code{\link{cav}} for more details
 \item \strong{cltau}: Refer to \code{\link{cltau}} for more details
 \item \strong{cltauw}: Refer to \code{\link{cltauw}} for more details
 \item \strong{ptr}: Refer to \code{\link{ptr}} for more details
 \item \strong{ptf}: Refer to \code{\link{ptf}} for more details
 \item \strong{vzo}: Refer to \code{\link{vzo}} for more details
 \item \strong{vzp}: Refer to \code{\link{vzp}} for more details
}
}
\section{Note}{

By default all the return list options are selected and calculated if 'parameter_list' is not specified. Please refer to \code{\link{mct_template}}
to get more calrification on how to specify which parameters to calculate to this function if you wish to subset the default caculated parameters. \cr
By default 'display_list' argument is empty, which means that this function will return all caculated parameters specifed by the 'parameter_list' argument.
Only specfiy a list of parameters to the 'display_list' if you want to subset the calculated parameters returned as a result of this function. \cr
By default 'return_list' argument is empty, which means that this function will not append parameters passed from 'data' argument.
Only specfiy a list of parameters to the 'return_list' if you want to return them as a result of this function. \cr
If 'optimize_kel' is FALSE AND KEL is not one of the parameters (default or specified by 'parameter_list' argument) then
this functions will return a datafram. If 'optimize_kel' is FALSE AND KEL is one of the parameters (default or specified by 'parameter_list' argument) then
this functions will return a list with 'data_out' and 'est_data'. If 'optimize_kel' is TRUE AND KEL is one of the parameters (default or specified
by 'parameter_list' argument) then this functions will return a list with 'data_out', 'optimized_kel_flag' and 'est_data'.
If 'optimize_kel' is TRUE AND KEL is not one of the parameters (default or specified by 'parameter_list' argument) then
this functions will return a list with 'data_out' and 'optimized_kel_flag'. \cr
Please note that this function does not contain all the features of a M2 SS computation, so it is recommended that you
use the parent function \code{\link{run_computation}}
}

\section{Returns}{

\strong{Dataframe} \cr
\itemize{
 \item M2SS_Parameters: Caculated default/specified parameters
}
OR \cr
\strong{List} \cr
\itemize{
 \item data_out: Calculated default/specified M2SS Parameters
 \item optimized_kel_flag: Optimized KEL flag data used to calulate KEL based parameters
 \item est_data: Calculated Estimated Parameters
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer & Rudraya Technical Team}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
