% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_gravity_adjustment.R
\name{specific_gravity_adjustment}
\alias{specific_gravity_adjustment}
\title{Adjust Urinary Amounts in weight units to volumne units}
\usage{
specific_gravity_adjustment(amt_units, specific_gravity = 1.02,
  map = NULL, verbose = FALSE)
}
\arguments{
\item{specific_gravity}{- human urine specific gravity, default to 1.020 g/mL}

\item{map}{Model Configuration Template (MCT/map) can include a user specified correct for specific
gravity or a data column that provides that information for a specific subject/experimental unit.}

\item{sample_amount}{The sample amount data (volume or weight) with one value/sample collection interval (given in a vector form)}

\item{sample_amountu}{The units for each sample amount measurement}

\item{na.rm}{Pass nm.rm to sum function, true by default}
}
\value{
ssga: specific gravity adjustment multiplier
}
\description{
Adjust Urinary Amounts in weight units to volumne units
}
\details{
If urine amounts are reported in weight units, apply specific gravity correction to convert to volumne units.
The standard correction utilized is 1.020 g/mL = specific gravity or urine. 

Default Human Urine Specific Gravity Value obtained from Pagana, KD. Mosbys
Diagnostic and Laboratory Test Reference (Eighth Edition). 2007, Page 968.
The range for adult is 1.005-1.030, with a median derived value of 1.020 g/mL.
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item sga: specific gravity adjustment multiplier
}
}

\examples{
specific_gravity_adjustment(amt_units="mL")
#[1] 1

specific_gravity_adjustment(amt_units="ml")
#[1] 1

specific_gravity_adjustment(amt_units="L")
#[1] 1

specific_gravity_adjustment(amt_units="mg")
#[1] 1.02

specific_gravity_adjustment(amt_units="GM")
#[1] 1.02

is.element("SPECIFIC_GRAVITY", names(mct))
#[1] FALSE
specific_gravity_adjustment(amt_units="GM", map=mct)
#[1] 1.02

mct$SPECIFIC_GRAVITY <- 1.03
specific_gravity_adjustment(amt_units="mL", map=mct)
#[1] 1
specific_gravity_adjustment(amt_units="gm", map=mct)
#[1] 1.03

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer, Inc}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
