% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlag.R
\name{tlag}
\alias{tlag}
\title{Lag Time}
\usage{
tlag(conc = NULL, time = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}
}
\description{
This function gets the lag time, which is the time before the start of absorption phase
and is directly observed from the concentration-time profile. This time is defined s the sample time
immediately prior to the first quantifiable concentration.
}
\details{
If all concentrations have a value greater than 0 then TLAG is 0.
You must provide a valid SDEID with respect to the data. \cr
The function will ensure that the concentration data and the time data provided are numeric vectors and
are also of the same length. \cr
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item TLAG: lag time
}
}

\examples{
##########
## Data ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   30    ##    0   ##   2.89   ##
##   30    ##    1   ##   2.49   ##
##   30    ##    2   ##   2.47   ##
###################################

#Data mentioned will be used for the following example

#tlag()
#Error in tlag: 'conc' and 'time' vectors are NULL

conc_vector <- c(2.89, 2.49, 2.47)
time_vector <- c(0, 1, 2)

tlag(conc = conc_vector, time = time_vector)
#0

############
## Data 2 ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   31    ##    0   ##      0   ##
##   31    ##    1   ##      0   ##
##   31    ##    2   ##      0   ##
###################################

#Data mentioned will be used for the following example

conc_vector <- c(0, 0, 0)
time_vector <- c(0, 1, 2)

tlag(conc = conc_vector, time = time_vector)
#NA

############
## Data 3 ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   32    ##    0   ##   NA     ##
##   32    ##    1   ##   1.23   ##
##   32    ##    2   ##   1.34   ##
##   32    ##    3   ##   1.19   ##
###################################

#Data mentioned will be used for the following example

conc_vector <- c(NA, 1.23, 1.34, 1.19)
time_vector <- c(0, 1, 2, 3)

tlag(conc = conc_vector, time = time_vector)
#1

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
