% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlast.R
\name{tlast}
\alias{tlast}
\title{Time at Last Measurable Plasma Concentration}
\usage{
tlast(data = NULL, map = NULL, flag = NULL, sid = list())
}
\arguments{
\item{data}{The model data that was generated by \code{\link{model}}}

\item{sid}{The subject ID (either can be the id number or 'all' to get tlast for all the subjects)}

\item{include}{To include zero's within the calculation for TLAST}
}
\description{
This function gets the time of last measurable plasma concentration that is obtained by the 
inspection of the data.
}
\details{
If all the concentrations are 0's then TLAST will return 'ND' for Not Determined. 
You must provide a valid sid with respect to the data.
}
\section{Returns}{

\strong{Dataset} \cr 
\itemize{
 \item SID: subject identification number
 \item TLAST: time at last mesuarable plasma concentration
}
}

\examples{
########## 
## Data ##
#################################
##  SID  ##  TIME  ##  RESULT  ##
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
##   31  ##    0   ##      0   ##
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
##   32  ##    0   ##   1.19   ##
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##      0   ##
##   32  ##    4   ##    BLQ   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

tlast()   
#No data found!

tlast(data)  
#Object not of class NCA

mod <- model("~/data.csv")  
#Creates an NCA object with data represented in 'data' above
clast(mod)  
#Please specify for which subject you want to get the TLAST for!

tlast(mod, sid = "all")  
# SID  TMAX
#  30     2
#  31    ND
#  32     2

tlast(mod, sid = 32)  
# SID  TMAX
#  32     2

tlast(mod, sid = 32, include = FALSE)  
# SID  TMAX
#  32     1

tlast(mod, sid = 10)  
Invaild subject ID! Subject ID not found in the data provided!

}
\author{
\itemize{
 \item Kevin McConnell
}
}
