% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmax.R
\name{tmax}
\alias{tmax}
\title{Time at Maximum Observed Concentration}
\usage{
tmax(conc = NULL, time = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}
}
\description{
This function gets the first time at which CMAXi is observed within a dosing interval 
and is obtained by the inspection of the data. In the case of multiple dosing, TMAXi is obtained
by inspection of the data during the dose inteval i.
}
\details{
If all the concentrations are 0's then TMAXi will return NA. 
Also the interval must be in the range of the times listed in the data.
}
\section{Returns}{

\strong{Value} \cr 
\itemize{
 \item TMAX: time at maximum observed concentration
}
}

\examples{
##########
## Data ##
#################################
##  SID  ##  TIME  ##   CONC   ## 
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

tmax()   
#Error in tmax: 'conc' and 'time' vectors are NULL

conc_vector <- data$CONC
time_vector <- data$TIME

tmax(conc = conc_vector, time = time_vector)
#0
 
############
## Data 2 ##
#################################
##  SID  ##  TIME  ##   CONC   ## 
#################################
##   31  ##    0   ##      0   ## 
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
#################################

data2 <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

conc_vector <- data2$CONC
time_vector <- data2$TIME

tmax(conc = conc_vector, time = time_vector)
#NA

############
## Data 3 ##
#################################
##  SID  ##  TIME  ##   CONC   ## 
#################################
##   32  ##    0   ##   1.19   ## 
##   32  ##    1   ##   1.27   ##
##   32  ##    2   ##   1.24   ##
##   32  ##    3   ##   1.27   ##
#################################

data3 <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

conc_vector <- data3$CONC
time_vector <- data3$TIME

tmax(conc = conc_vector, time = time_vector)
#1

}
\author{
\itemize{
 \item Kevin McConnell
}
}
