% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmax_i.R
\name{tmax_i}
\alias{tmax_i}
\title{Time at Maximum Observed Concentration}
\usage{
tmax_i(data = NULL, map = NULL, flag = NULL, sid = list(),
  interval = 0)
}
\arguments{
\item{data}{The model data that was generated by \code{\link{model}}}

\item{sid}{The subject ID (either can be the id number or 'all' to get tmax for all the subjects)}

\item{interval}{The endpoint of an dosing interval (if not provided then it will use the entire time interval)}
}
\description{
This function gets the first time at which CMAXi is observed within a dosing interval 
and is obtained by the inspection of the data. In the case of multiple dosing, TMAXi is obtained
by inspection of the data during the dose inteval i.
}
\details{
If all the concentrations are 0's then TMAXi will return 'ND' for Not Determined. 
Also the interval must be in the range of the times listed in the data. 
And also you must provide a valid sid with respect to the data.
}
\section{Returns}{

\strong{Dataset} \cr 
\itemize{
 \item SID: subject identification number
 \item TMAX: time at maximum observed concentration
 \item INTERVAL: endpoint of the dosing interval
}
}

\examples{
##########
## Data ##
#################################
##  SID  ##  TIME  ##  RESULT  ##
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
##   31  ##    0   ##      0   ##
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
##   32  ##    0   ##   1.19   ##
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##   1.34   ##
##   32  ##    4   ##   1.32   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

tmax_i()   
#No data found!

tmax_i(data)  
#Object not of class NCA

mod <- model("~/data.csv")  
#Creates an NCA object with data represented in 'data' above
tmax_i(mod)  
#Please specify for which subject you want to get the TMAXi for!

tmax_i(mod, sid = "all")  
# SID  TMAX  INTERVAL
#  30     0         0
#  31    ND         0
#  32     2         0

tmax_i(mod, sid = 31)  
# SID  TMAX  INTERVAL
#  31    ND         0

tmax_i(mod, sid = 10)  
#Invaild subject ID! Subject ID not found in the data provided!

tmax_i(mod, sid = 31, interval = -1)
#Interval provided is less than zero!

tmax_i(mod, sid = 31, interval = 4)
#Interval provided is greater than maximum time!

tmax_i(mod, sid = 32, interval = 1)
# SID  TMAX  INTERVAL
#  32     1         1

tmax_i(mod, sid = "all", interval = 4)
#Interval provided is greater than maximum time!

}
\author{
\itemize{
 \item Kevin McConnell
}
}
