% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmaxrate.R
\name{tmaxrate}
\alias{tmaxrate}
\title{Midpoint of collection interval associated with the maximum observed excretion rate}
\usage{
tmaxrate(midpt = NULL, rate = NULL)
}
\arguments{
\item{midpt}{The midpoint data (given in a vector form)}

\item{rate}{The rate data (given in a vector form)}
}
\description{
Midpoint of collection interval associated with the maximum observed excretion rate
}
\details{
If all the RATE's are 0's then RATELAST will return 'NA'.
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item TMAXRATE: maximum observed excretion rate
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
