% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmin.R
\name{tmin}
\alias{tmin}
\title{Time at Minimum Observed Concentration}
\usage{
tmin(conc = NULL, time = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}
}
\description{
This function gets the first time at which minimum observed concentration is observed
within the dosing interval and is obtained by the inspection of the data. In the case of
multiple dosing, TMINi is obtained by inspection of the data during each dosing interval i.
}
\details{
If all the concentrations are 0's then TMINi will return NA. \cr
The function will ensure that the concentration data and the time data provided are numeric vectors and
are also of the same length. \cr
}
\section{Returns}{

\strong{Value} \cr
\itemize{
 \item TMIN: time at minimum observed concentration
}
}

\examples{
##########
## Data ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   30    ##    0   ##   2.89   ##
##   30    ##    1   ##   2.49   ##
##   30    ##    2   ##   2.47   ##
###################################

#Data mentioned will be used for the following example

tmin()
#Error in tmin: 'conc' and 'time' vectors are NULL

conc_vector <- c(2.89, 2.49, 2.47)
time_vector <- c(0, 1, 2)

tmin(conc = conc_vector, time = time_vector)
#2

############
## Data 2 ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   31    ##    0   ##      0   ##
##   31    ##    1   ##      0   ##
##   31    ##    2   ##      0   ##
###################################

#Data mentioned will be used for the following example

conc_vector <- c(0, 0, 0)
time_vector <- c(0, 1, 2)

tmin(conc = conc_vector, time = time_vector)
#NA

############
## Data 3 ##
###################################
##  SDEID  ##  TIME  ##   CONC   ##
###################################
##   32    ##    0   ##   1.19   ##
##   32    ##    1   ##   1.23   ##
##   32    ##    2   ##   1.34   ##
##   32    ##    3   ##   1.19   ##
###################################

#Data mentioned will be used for the following example

conc_vector <- c(1.19, 1.23, 1.34, 1.19)
time_vector <- c(0, 1, 2, 3)

tmin(conc = conc_vector, time = time_vector)
#0

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
