% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmin_i.R
\name{tmin_i}
\alias{tmin_i}
\title{Time at Minimum Observed Concentration}
\usage{
tmin_i(data = NULL, map = NULL, flag = NULL, sid = list(),
  interval = 0)
}
\arguments{
\item{data}{The model data that was generated by \code{\link{model}}}

\item{sid}{The subject ID (either can be the id number or 'all' to get tmin for all the subjects)}

\item{interval}{The endpoint of an dosing interval (if not provided then it will use the entire time interval)}
}
\description{
This function gets the first time at which CMINi is observed within the dosing interval and
is obtained by the inspection of the data. In the case of multiple dosing, TMINi 
is obtained by inspection of the data during each dosing interval i.
}
\details{
If all the concentrations are 0's then TMINi will return 'ND' for Not Determined. 
Also if the interval provided is less than 0 then TMINi will return 'ND' for Not Determined. 
And also you must provide a valid sid with respect to the data.
}
\section{Returns}{

\strong{Dataset} \cr 
\itemize{
 \item SID: subject identification number
 \item TMIN: time at minimum observed concentration
 \item INTERVAL: endpoint of the dosing interval
}
}

\examples{
##########
## Data ##
#################################
##  SID  ##  TIME  ##  RESULT  ##
#################################
##   30  ##    0   ##   2.89   ##
##   30  ##    1   ##   2.49   ##
##   30  ##    2   ##   2.47   ##
##   31  ##    0   ##      0   ##
##   31  ##    1   ##      0   ##
##   31  ##    2   ##      0   ##
##   32  ##    0   ##   1.19   ##
##   32  ##    1   ##   1.23   ##
##   32  ##    2   ##   1.34   ##
##   32  ##    4   ##   1.32   ##
#################################

data <- data.frame(
    SID = ...,
    TIME = ...,
    RESULT = ...
)
#Same data as above, just represented as a dataframe

tmin_i()   
#No data found!

tmin_i(data)  
#Object not of class NCA

mod <- model("~/data.csv")  
#Creates an NCA object with data represented in 'data' above
tmin_i(mod)  
#Please specify for which subject you want to get the TMINi for!

tmin_i(mod, sid = "all")  
# SID  TMAX  INTERVAL
#  30     2         0
#  31    ND         0
#  32     0         0

tmin_i(mod, sid = 31)  
# SID  TMAX  INTERVAL
#  31    ND         0

tmin_i(mod, sid = 10)  
#Invaild subject ID! Subject ID not found in the data provided!

tmin_i(mod, sid = 31, interval = -1)
#Interval provided is less than zero!

tmin_i(mod, sid = 31, interval = 4)
#Interval provided is greater than maximum time!

tmin_i(mod, sid = 30, interval = 1)
# SID  TMAX  INTERVAL
#  30     1         1

tmin_i(mod, sid = "all", interval = 4)
#Interval provided is greater than maximum time!

}
\author{
\itemize{
 \item Kevin McConnell
}
}
