% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Run Computation}
\usage{
unit_conversion(data = NULL, map = NULL, result = NULL,
  unit_class = "ALL", verbose = TRUE)
}
\arguments{
\item{data}{The dataframe that contains the raw data}

\item{map}{The dataframe that contains the map data}

\item{result}{The dataframe that contains the computation engine result data}
}
\description{
This function will compute all the relevant parameters for based specified fields provided via map argument.\cr
}
\details{
Please refer to \code{\link{mct_template}} for more details on what the 'map' argument dataframe should contain.
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item data_out: Calculated default/specified M1SS Parameters
 \item result_data: Result Data used during calculation of parameters
 \item est_data: Calculated Estimated Parameters
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Tom Tensfeldt, Pfizer Pharmacometrics & Rudraya Technical Team}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
