% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitclass_parameters.R
\name{unitclass_parameters}
\alias{unitclass_parameters}
\title{unitclass_parameters}
\usage{
unitclass_parameters(parameter, dl = create_dependency_list())
}
\arguments{
\item{parameter}{Parameter}

\item{dl}{Dependency List (Please refer to \code{\link{create_dependency_list}})}
}
\description{
This function returns the list of parameters based on the unit class specified
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item List of Parameters based on Unit Class
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
