% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mct_data.R
\name{update_mct_data}
\alias{update_mct_data}
\title{update_mct_data updates time, alternative time and concentration data}
\usage{
update_mct_data(map, data, flag, verbose = FALSE)
}
\arguments{
\item{map}{Model Configuration Template (MCT)}

\item{data}{input concentration dataset}

\item{verbose}{logs selected time/concentration values and units to stdout}

\item{flags}{input flags dataset}
}
\value{
map
}
\description{
update_mct_data executes validate_timeconc_data and updates Model Configuration Template/MCT/map values to align with
selected TIME, alternative TIME, CONC and associated units.
}
\details{
that map$TIME variable either points to a valid variable in data or is nominal/actual selection.
map$TIME is either a value of "Nominal" or "Actual" or it points directly to a column/data field in the data dataframe.
if map$TIME=="Nominal" then map$NOMTIME/map$NOMTIMEU will point to the column/data field in the data dataframe for TIME
if map$TIME=="Actual"  then map$ACTTIME/map$ACTTIMEU will point to the column/data field in the data dataframe for TIME
if map$TIME points directly to a column/data field in the data dataframe, then map$TIMEU is an required field in map that
 points to avalud column/data field in the data dataframe for the units of time.
}
\section{Returns}{

updated Model Confuguration Template/MCT/map
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
