% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_models.R
\name{valid_models}
\alias{valid_models}
\title{valid_models}
\usage{
valid_models(dl = create_dependency_list())
}
\arguments{
\item{dl}{Dependency List (Please refer to \code{\link{create_dependency_list}})}
}
\description{
This function returns the list of valid model configurations that are supported
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item List of Valid Model configurations
}
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
