% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_timeconc_data.R
\name{validate_timeconc_data}
\alias{validate_timeconc_data}
\title{validate_timeconc_data validates that time, concentration, and dose data were passed via data and map/MCT and are available}
\usage{
validate_timeconc_data(map, data, verbose = FALSE)
}
\arguments{
\item{map}{Model Configuration Template (MCT)}

\item{data}{input concentration dataset}

\item{verbose}{logs selected time/concentration values and units to stdout}
}
\description{
validate_timeconc_data checks that map$TIME variable either points to a valid variable in data or is nominal/actual selection.
map$TIME is either a value of "Nominal" or "Actual" or it points directly to a column/data field in the data dataframe.
if map$TIME=="Nominal" then map$NOMTIME/map$NOMTIMEU will point to the column/data field in the data dataframe for TIME
if map$TIME=="Actual"  then map$ACTTIME/map$ACTTIMEU will point to the column/data field in the data dataframe for TIME
if map$TIME points directly to a column/data field in the data dataframe, then map$TIMEU is an required field in map that
 points to avalud column/data field in the data dataframe for the units of time.
}
\section{Returns}{

list with names time, timeu, conc, concu
}

\examples{

}
\author{
\itemize{
 \item \strong{Thomas Tensfeldt, Pfizer}
 \item website: \url{www.pfizer.com}
 \item email: \url{thomas.g.tensfeldt@pfizer.com}
}
}
