% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_lin.R
\name{auc_lin}
\alias{auc_lin}
\title{Linear Trapezoidal Area Under the Concentration-Time Curve (AUC) Calculation Method}
\usage{
auc_lin(conc = NULL, time = NULL, exflag = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{exflag}{The exclude flag data (given in a numeric vector)}
}
\description{
The AUC linear method is used for the entire profile.
}
