% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_lin_log.R
\name{auc_lin_log}
\alias{auc_lin_log}
\title{Linear-Log Trapezoidal Area Under the Concentration-Time Curve (AUC) Calculation Method}
\usage{
auc_lin_log(conc = NULL, time = NULL, exflag = NULL, t_max = NULL,
  interpolate = NULL, model = NULL, dosing_type = NULL,
  told = NULL, orig_conc = NULL, orig_time = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{exflag}{The exclude flag data (given in a numeric vector)}

\item{t_max}{The first time at which CMAXi is observed within the dosing interval (numeric value)}

\item{interpolate}{The value to determine whether to interpolate data points (given in a logical form)}

\item{model}{The model specification (either 'M1', 'M2', 'M3', or 'M4')}

\item{dosing_type}{The dosing type specification (either 'SD' or 'SS')}

\item{told}{The time of last dose (given in a numeric value)}

\item{orig_conc}{The original (full) concentration data (given in a numeric vector)}

\item{orig_time}{The original (full) time data (given in a numeric vector)}
}
\description{
The Linear-Log Trapezoidal Rule is applied with this AUC method. The linear method is used
up to Tmax (the first occurrence of Cmax), and the log trapezoidal method is used for the
remainder of the profile. If Ci or Ci+1 is 0 then the linear trapezoidal rule is used.
}
