% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_lin_log.R
\name{auc_lin_log}
\alias{auc_lin_log}
\title{Linear-Log Trapezoidal Area Under the Concentration-Time Curve (AUC) Calculation Method}
\usage{
auc_lin_log(conc = NULL, time = NULL, exflag = NULL, t_max = NULL)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{exflag}{The exclude flag data (given in a numeric vector)}

\item{t_max}{The first time at which CMAXi is observed within the dosing interval (numeric value)}
}
\description{
The Linear-Log Trapezoidal Rule is applied with this AUC method. The linear method is used
up to Tmax (the first occurrence of Cmax), and the log trapezoidal method is used for the
remainder of the profile. If Ci or Ci+1 is 0 then the linear trapezoidal rule is used.
}
