% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dof.R
\name{dof}
\alias{dof}
\title{Duration of infusion for constant infusion model M3}
\usage{
dof(data, map, dof_name = "DOF1")
}
\arguments{
\item{data}{The dataframe that contains the raw data}

\item{map}{The dataframe that contains the map data}

\item{dof_name}{The name of the DOF column in the map data}
}
\description{
This function returns the duration of infusion
}
\section{Returns}{

\strong{List} \cr
\itemize{
 \item DOF: duration of infusion
}
}

\examples{
#No appropriate examples

}
\author{
\itemize{
 \item \strong{Rudraya Technical Team}
 \item website: \url{www.rudraya.com}
 \item email: \url{support@rudraya.com}
}
}
