% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_told_concentration.R
\name{estimate_told_concentration}
\alias{estimate_told_concentration}
\title{Helper Function for Estimating TOLD Concentration Data}
\usage{
estimate_told_concentration(
  conc = NULL,
  time = NULL,
  interpolate = NULL,
  extrapolate = NULL,
  auc_method = NULL,
  model = NULL,
  dosing_type = NULL,
  told = NULL,
  orig_conc = NULL,
  orig_time = NULL,
  tmp = NULL
)
}
\arguments{
\item{conc}{The concentration data (given in a vector form)}

\item{time}{The time data (given in a vector form)}

\item{interpolate}{The value to determine whether to interpolate data points (given in a logical form)}

\item{extrapolate}{The value to determine whether to extrapolate data points (given in a logical form)}

\item{auc_method}{The method specified for AUC (either 'LIN' or 'LOG')}

\item{model}{The model specification (either 'M1', 'M2', 'M3', or 'M4')}

\item{dosing_type}{The dosing type specification (either 'SD' or 'SS')}

\item{told}{The time of last dose (given in a numeric value)}

\item{orig_conc}{The original (full) concentration data (given in a numeric vector)}

\item{orig_time}{The original (full) time data (given in a numeric vector)}
}
\description{
Helper Function for Estimating TOLD Concentration Data
}
