% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_lin.R
\name{interpolate_lin}
\alias{interpolate_lin}
\title{interpolate_lin performs linear interpolation between two data points}
\usage{
interpolate_lin(
  conc1 = NULL,
  time1 = NULL,
  conc2 = NULL,
  time2 = NULL,
  est_time = NULL
)
}
\description{
interpolate_lin performs linear interpolation between two data points
}
