% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.R
\name{rate}
\alias{rate}
\title{Excretion Rate for each collection interval}
\usage{
rate(
  start_time = NULL,
  end_time = NULL,
  conc = NULL,
  vol = NULL,
  volu = NULL,
  type = "INTERVAL",
  map = NULL
)
}
\description{
Excretion Rate for each collection interval
}
